refactor: move cobra commands into commands package
